Check if the Applications directory exists
The application packaging step will be skipped when the Applications
folder does not exist. For repositories that do not build
applications it makes sense to delete the Applications directory so
that the repo is less bloated. Since the Applications folder exists
by default there needs to be a explicit action to remove the folder
and therefore this change should not cause any silent failures
(applications now being built).